+++ src/App.tsx
@@ -1,4 +1,4 @@
-import { AppRoutes } from './routes';
+import { AppRoutes } from '@/routes';
 
 function App() {
   return <AppRoutes />;
