+++ src/components/Layout/PageLayout/index.ts
@@ -1,0 +1,1 @@
+export * from './PageLayout';
