+++ src/components/Layout/PageLayout/parts/Sidebar.tsx
@@ -1,5 +1,4 @@
-import { Logo } from './Logo';
-import { SideNavigation } from './SideNavigation';
+import { Logo, SideNavigation } from '.';
 
 export const Sidebar = () => {
   return (
