+++ src/components/Layout/PageLayout/parts/index.ts
@@ -1,0 +1,4 @@
+export * from './Logo';
+export * from './Sidebar';
+export * from './SideNavigation';
+export * from './Topbar';
